refactor : network 모듈 분리
