+++ network/src/main/java/com/kky/pokedex/network/model/Info.kt
@@ -1,4 +1,4 @@
-package com.kky.pokedex.model
+package com.kky.pokedex.network.model
 
 data class Info(
     val name: String,
